moves getImage and getUser functions
